rules/requirements.txt
CherryPy==18.1.0
ws4py==0.5.1
PyJWT==2.0.1
Pypubsub==4.0.3
websocket-client==0.57.0